credit-usage-function/requirements.txt
psycopg2-binary
google-cloud-pubsub
google-api-core